fixed bug in HMM training unittests
Old test for increasing log likelihood would pass as long as the log
likehood did not *decrease* by more than 1000.  Moved the threshold up
to 0.5 (not 0 because there are sometimes numerical issues, etc.)

Now that the test was fixed, it turns out that the training code for
'tied' and 'full' covariances are still buggy.